type consistency with tf

+++ cache_numpy.py
@@ -22,7 +22,7 @@ class LRUCache:
     def QueryMultipleEntries(self, matching_vectors):
         num_entries = matching_vectors.shape[1]
         matching_vectors = matching_vectors.reshape(num_entries, -1)
-        results = np.zeros_like(matching_vectors)
+        results = np.zeros_like(matching_vectors, dtype=np.float32)
 
         for i in range(num_entries):
             results[i] = self.Query(matching_vectors[i])
